pyproject.toml
[tool.poetry]
name = "pyaioconsole"
version = "0.1.1"
description = "Python library for creating async CLI applications"
authors = ["alexeev-prog <alexeev.dev@mail.ru>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
rich = "^13.9.4"
asyncio = "^3.4.3"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "B", "I"]
ignore = []
fixable = ["ALL"]

[tool.ruff.lint.mccabe]
max-complexity = 5

[tool.ruff.lint.isort]
force-single-line = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
